+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to
 - Add support for new tools.
 - Add packages checking support for Arch Linux.
 
+## [1.1.0] - 2019-04-27
+### Added
+- Add a .deb package for easy dependencies install.
+- Add alternative packages names support.
+### Changed
+- Ignore comments and empty lines of the patches.
+### Fixed
+- Correct associative arrays declaring.
+
 ## [1.0.1] - 2019-04-25
 ### Fixed
 - Change an output directory for binaries.